requirements.txt
snowflake-snowpark-python[pandas]
sklearn
pandas
pytest